requirements.txt
# Web Framework
fastapi
uvicorn

# AI Model
ollama

# Speech to Text
openai-whisper
sounddevice
scipy
numpy

# Text to Speech
pyttsx3

# Audio
soundfile

# PyTorch with CUDA for GPU
torch
torchaudio